Add game information in welcome modal
